requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
streamlit>=1.28.0
